Split up code into separate modules
