+++ crates/cef-ui-util/src/builds/link_cef.rs
@@ -1,4 +1,4 @@
-use crate::{copy_files, download_and_extract_cef, get_build_rs_target_dir};
+use crate::download_and_extract_cef;
 use anyhow::Result;
 use std::path::Path;
