+++ crates/cef-ui-util/src/lib.rs
@@ -1,7 +1,7 @@
-mod link;
+mod builds;
 mod tools;
 mod util;
 
-pub use link::*;
+pub use builds::*;
 pub use tools::*;
 pub use util::*;
