+++ crates/cef-ui-util/src/tools/cef_build.rs
@@ -1,11 +1,6 @@
-use crate::{copy_files, get_tool_target_dir};
 use anyhow::Result;
 use clap::Parser;
-use std::{
-    env::current_dir,
-    fs::{copy, create_dir_all, remove_dir_all},
-    process::{Command, Stdio}
-};
+use std::process::{Command, Stdio};
 use tracing::{info, level_filters::LevelFilter, subscriber::set_global_default, Level};
 use tracing_log::LogTracer;
 use tracing_subscriber::FmtSubscriber;
@@ -40,6 +35,8 @@ pub fn cef_build() -> Result<()> {
     // On macOS, build the helper executable
     // and package the app bundle as required.
     if cfg!(target_os = "macos") {
+        use crate::builds::build_app_bundle;
+
         info!("Building helper binary ..");
 
         build_exe("cef-ui-simple-helper", &args.profile)?;
@@ -53,7 +50,7 @@ pub fn cef_build() -> Result<()> {
 }
 
 /// Build a specific executable.
-fn build_exe(name: &str, profile: &str) -> Result<()> {
+pub fn build_exe(name: &str, profile: &str) -> Result<()> {
     let args = vec!["build", "--bin", name, "--profile", profile];
 
     Command::new("cargo")
@@ -61,94 +58,6 @@ fn build_exe(name: &str, profile: &str) -> Result<()> {
         .stdout(Stdio::inherit())
         .stderr(Stdio::inherit())
         .output()?;
-
-    Ok(())
-}
-
-/// Package the app bundle on macOS.
-fn build_app_bundle(profile: &str) -> Result<()> {
-    let cwd = current_dir()?;
-    let target_dir = get_tool_target_dir(profile)?;
-    let app_dir = target_dir.join("cef-ui-simple.app");
-    let resources_dir = cwd.join("resources/macos");
-
-    // Remove any existing app.
-    if app_dir.exists() {
-        remove_dir_all(&app_dir)?;
-    }
-
-    // Create main bundle folders.
-    create_dir_all(app_dir.clone())?;
-    create_dir_all(app_dir.join("Contents/Frameworks"))?;
-    create_dir_all(app_dir.join("Contents/MacOS"))?;
-    create_dir_all(app_dir.join("Contents/Resources"))?;
-
-    // Copy main bundle files.
-    copy(
-        resources_dir.join("Info.plist"),
-        app_dir.join("Contents/Info.plist")
-    )?;
-
-    copy(
-        resources_dir.join("Icon.icns"),
-        app_dir.join("Contents/Resources/Icon.icns")
-    )?;
-
-    copy_files(
-        &resources_dir.join("English.lproj"),
-        &app_dir.join("Contents/Resources/English.lproj")
-    )?;
-
-    copy(
-        target_dir.join("cef-ui-simple"),
-        app_dir.join("Contents/MacOS/cef-ui-simple")
-    )?;
-
-    // Copy the CEF framework.
-    copy_files(
-        &cwd.join("artifacts/cef/Chromium Embedded Framework.framework"),
-        &app_dir.join("Contents/Frameworks/Chromium Embedded Framework.framework")
-    )?;
-
-    let create_helper = |name: Option<&str>| -> Result<()> {
-        let helper_name = match name {
-            Some(name) => format!("cef-ui-simple Helper ({})", name),
-            None => "cef-ui-simple Helper".to_string()
-        };
-
-        let helper_dir = app_dir.join(format!("Contents/Frameworks/{}.app", helper_name));
-
-        // Create helper bundle folders.
-        create_dir_all(helper_dir.clone())?;
-        create_dir_all(helper_dir.join("Contents/MacOS"))?;
-
-        // Copy helper bundle files.
-        let plist_name = match name {
-            Some(name) => format!("{}HelperInfo.plist", name),
-            None => "HelperInfo.plist".to_string()
-        };
-
-        copy(
-            resources_dir.join(plist_name),
-            helper_dir.join("Contents/Info.plist")
-        )?;
-
-        copy(
-            target_dir.join("cef-ui-simple-helper"),
-            helper_dir
-                .join("Contents/MacOS")
-                .join(helper_name)
-        )?;
-
-        Ok(())
-    };
-
-    // Create the helper bundles.
-    create_helper(None)?;
-    create_helper(Some("Alerts"))?;
-    create_helper(Some("GPU"))?;
-    create_helper(Some("Plugin"))?;
-    create_helper(Some("Renderer"))?;
 
     Ok(())
 }
